+++ geotrek/api/tests/test_mobile/test_sync_mobile.py
@@ -389,6 +389,14 @@ class SyncMobileTreksTest(TranslationResetMixin, VarTmpTestCase):
                                                         published=True)
         cls.touristic_event = TouristicEventFactory(geom='SRID=%s;POINT(700001 6600001)' % settings.SRID,
                                                     published=True)
+        cls.touristic_content_portal_a = TouristicContentFactory(geom='SRID=%s;POINT(700001 6600001)' % settings.SRID,
+                                                                 published=True, portals=[cls.portal_a])
+        cls.touristic_event_portal_a = TouristicEventFactory(geom='SRID=%s;POINT(700001 6600001)' % settings.SRID,
+                                                             published=True, portals=[cls.portal_a])
+        cls.touristic_content_portal_b = TouristicContentFactory(geom='SRID=%s;POINT(700001 6600001)' % settings.SRID,
+                                                                 published=True, portals=[cls.portal_b])
+        cls.touristic_event_portal_b = TouristicEventFactory(geom='SRID=%s;POINT(700001 6600001)' % settings.SRID,
+                                                             published=True, portals=[cls.portal_b])
         cls.attachment_content_1 = AttachmentFactory.create(content_object=cls.touristic_content,
                                                             attachment_file=get_dummy_uploaded_image())
         cls.attachment_event_1 = AttachmentFactory.create(content_object=cls.touristic_event,
@@ -424,14 +432,22 @@ class SyncMobileTreksTest(TranslationResetMixin, VarTmpTestCase):
         management.call_command('sync_mobile', 'var/tmp', url='http://localhost:8000',
                                 skip_tiles=True, verbosity=2, portal=self.portal_a.name, stdout=output)
         self.assertFalse(os.path.exists(
-            os.path.join('var/tmp/en', '{pk}'.format(pk=str(self.trek_3.pk)), 'trek.geojson')
+            os.path.join('var', 'tmp', 'en', str(self.trek_3.pk), 'trek.geojson')
         ))
         for lang in settings.MODELTRANSLATION_LANGUAGES:
-            with open(os.path.join('var/tmp', lang, 'treks.geojson'), 'r') as f:
+            with open(os.path.join('var', 'tmp', lang, 'treks.geojson'), 'r') as f:
                 trek_geojson = json.load(f)
                 self.assertEqual(len(trek_geojson['features']),
                                  Trek.objects.filter(**{'published_{}'.format(lang): True})
                                  .filter(Q(portal__name__in=(self.portal_a,)) | Q(portal=None)).count())
+        with open(os.path.join('var', 'tmp', 'en', str(self.trek_1.pk), 'touristic_contents.geojson'), 'r') as f:
+            tc_geojson = json.load(f)
+            self.assertEqual(len(tc_geojson['features']), 1)
+            # Only one because factory generate a portal for touristic contents
+        with open(os.path.join('var', 'tmp', 'en', str(self.trek_1.pk), 'touristic_events.geojson'), 'r') as f:
+            te_geojson = json.load(f)
+            # Two because factory do not generate a portal for touristic events
+            self.assertEqual(len(te_geojson['features']), 2)
 
     def test_sync_pois_by_treks(self):
         output = StringIO()