+++ docs/changelog.rst
@@ -10,6 +10,10 @@ CHANGELOG
 - Fix bootstrap theme in warning and error messages or alerts
 - Fix Services external IDs were not displayed in detail pages
 
+**Minor improvements**
+
+- Add filtering portals sync_mobile for touristic contents and events
+
 
 2.74.1     (2021-12-21)
 -----------------------